feat: implement sequence helper function in array
TypeScript still has issues inferring the correct type for:

      import * as A from "https://deno.land/x/fun/array.ts";
      import * as O from "https://deno.land/x/fun/option.ts";
      import { identity, pipe } from "https://deno.land/x/fun/fns.ts";

      const sequence = pipe(
        O.map(identity),
        A.traverse(O.Applicative)
      )

Since there is a call for an easy sequence (not over a tuple) of arrays
of various ADTs this commit implements a createSequence function in
array.ts.

      import * as A from "../array.ts";
      import * as O from "../option.ts";
      import { identity, pipe } from "../fns.ts";

      const sequence = A.createSequence(O.Applicative);

      console.log(sequence([O.some(1), O.some(2)])); // { tag: "Some", value: [ 1, 2 ] }
      console.log(sequence([O.none, O.some(2)])); // { tag: "None" }

While it is not used to create an adt specific sequence function for
every ADT it's possible that it might be used thusly in the future.