fix: correct tests for hoisted functions
It has become apparent that the deno coverage generation has some
flaws. The coverage reporter finds some lines to be uncovered even
when they have multiple tests covering all branches. For this reason
we are going to put test completion on hold until the tools for
observing coverage are complete.