feat: schemable rewrite
* hoist funcs for schemable.ts
* hoist funcs for decoder.ts
* hoist funcs for guard.ts
* migrate and hoist funcs for json.ts

For the original port of schemable to fun I focused on simplifying the
implementation. io-ts uses a Kliesli specific to decoder in order to
make implementation of decoder-like Schemables easier. In the process
this makes the standard decoder incredibly difficult to reason about and
doesn't seem to actually reduce amount of code used. Unfortunately, the
decoder type was neutered during this process, making decoders/guards
specific to parsing strings to dates and general refinement impossible.

To this end, this commit reintroduces the in and out types for Guard
and Decoder. Each schemable decoder is implemented as a root function
and the Schemable instances are built from these (instead of building
the individual sub-schemables).

The DecodeError type was also simplified. I tried using Tree directly,
along with a custom data-type coupled with a manual Semigroup.
Ultimately, I came back to the Free semigroup but implemented the entire
Result type as a Right Monad over Either. The original DecodeError type
is still usable but the Result type is the primary export.

Lastly, this commit also pulls the jsonschema library into the json.ts
file under schemable. In general this is because I want fun to operate
more like a fully fledged functional library. This will help keep the
maintenance burden light in future changes. I'm not certain where the
line will be drawn but the next set of features is sure to include a
stream implementation, a parser, and other tools. Unlike fp-ts it seems
more prudent to allow contributions to have space within fun rather than
shunting them to something like fp-ts-contrib.

FossilOrigin-Name: 2b61a254ba51de29795bf237d2234406ce416ca8b26c87ab073b18bb892f07f5