feat: pivot to iterable and async iterable
After some difficult work on a modified callbag implementation for
streams I was introduced to the IxJS AsyncIterable as well as the WHATWG
Streams spec. It seems prudent to implement fun modules over these
native data types before going full de-novo. Thus, there are new modules
for Iterable as well as AsyncIterable.

NOTE: These datatypes are not immutable by default but there is no way to
type the Mutability of either data type.

A few additions were made to the optics libraries as well.