feat: initial port of newtype-ts
This is a minimal port of the fp-ts ecosystem newtype. Newtype is
primarily used to narrow a broad type (like number) to something specific
(like USD or EUR) and to limit compile time use to the narrow type (ie.
a function that takes USD would show a compile time error if it was
passed EUR, even though both narrow types are implemented with `number`).