fix: correct type for schemable > nullable
The NullableSchemable type in schemable.ts had set the second generic
in its Kind return value to `B | null`. This type is incorrect as it
assumes that the second parameter of the replacement type must be null.
For example, NullableSchemable for a Decoder might have the type
`type numToStrDecode = Decoder<number, string>`. Applying the nullable
schemable to this decoder would result in the erroneous type
`Decoder<number | null, string | null>`. The fix is to remove the union
with null from the output second parameter of NullableSchemable.

Fixes #26