feat: move at.ts implementation to lens.ts
* move and refactor or atRecord and atMap
* atMap now takes a setoid since it can have more than string keys
* removed indexed algebraic structures since they were not
  used anywhere. We get the benefits of indexing by implementing
  it at the combinator level and those combinators still work
  for the non-indexed algebraic structures.
* updated testing for all changes
* modified the type for record to track known keys. this has
  nice benefits for typings but has a rough edge case around
  optional parameters when the type checker doesn't have exact
  optional properties enabled (the default). Basically, don't
  use named optional properties if you can help it, as they
  can have existing keys but be set to undefined, breaking
  the record.ts logic.
* updated all dependencies on atRecord and atMap

Fixes #70

feat: move index.ts implementation to optional.ts

* undid type changes to record.ts :D
* moved indexArray and indexRecord to optional.ts
* expanded typeof of indexRecord to use index type
* implemented indexMap in optional.ts
* updated dependencies on indexArray and indexRecord

This ended up forcing some changes for the insert, update, and modify
functions across map, record, and array. From now on, if an ADT is
a collection type with an index the following combinators need to be
created (note that some indexes require a setoid for comparison
but some do not. In the case of array and record, the setoid is
omitted):

* lookup: setoid => key => collection => Option<value>
* delete: setoid => key => collection => collection

Insert will always insert if possible, regardless of the existence
of an existing value at the key. Think of it like upsert. For array
insert can fail if the key/index is out of bounds, thus it returns
Option<collection>. For array, insert always extends the array by
one value. This means that insert for array can be done at the end
of the array. For other collections like record and map, a value
is replaced or added.

* insert: setoid => value => key => collection => collection
* insertAt: setoid => key => value => collection => collection

Modify will only modify a value if it already exists.

* modify: setoid => modifyFn => key => collection => collection
* modifyAt: setoid => key => modifyFn => collection => collection

Update will only change a value if it already exists.

* modify: setoid => value => key => collection => collection
* modifyAt: setoid => key => value => collection => collection

These functions have all been created and standardized across
array.ts, record.ts, and map.ts.

* fns.ts: added has function wrapping Object.hasOwn.
* index.ts: deleted
* optional.ts: added indexArray, indexRecord, and indexMap
* record.ts: added functions listed above and settle on string keys
* array.ts: standardized functions listed above
* map.ts: standardized functions listed above
* lens.ts: simplified atRecord and atMap implementations

Fixes #70

FossilOrigin-Name: 5ee9b73ebc30f61c46a3b841b48882089ef27c011c8301feafb184089b052e01