feat: json_schema 2
Renamed parts of JsonSchema to be more clear. Any types that
represent literal JSON Schema types were prepended with
JsonSchema (eg. String => JsonSchemaString).

This led to renaming JsonSchema to JsonBuilder, which is more
descriptive of what the type is used for.

I added a Hold type to Kind and used it to implement the
TypeClass type. These are marked as experiemental until
I understand their uses. I've also added TypeClass to
all of the "Type Classes" in fun

Updated all documentation and tests for json schema.

Lastly, I altered json_schema.ts print to not output
definitions if there are none to print.