feat: remove types.ts and breakout all types
The types.ts file was meant first as a catch all for non-adt
or typeclass types. At some point it also because a type only
barrel export.

In the pursuit of more explicit imports it has become obvious that
the types.ts file is no longer necessary, and in fact becomes
an issue in some edge cases.

Thus, it has been removed in favor of the explicit dependency
graphs generated by avoiding all barrel exports.

Fixes #163