feat: implement ord 2.0
While it 'might' be possible to implement most of schemable for
Ord, I've opted to punt that to a maybe later date. On the other
hand, creating adhoc max, min, clamp, etc functions from an ord
is a pain. So for now I've decided to implement all of the
comparison operations directly on Ord. Almost all ords are
easier to create with the fromCompare constructor anyway.

* wrote the file header
* implemented all combinators and derivations
* modified all functions to adhere to new style guide
* documented every export
* upped test coverage to 100%

Fixes #108