feat: switch to pragmatic type class names
This is a big dumb commit where I dropped all of the Category Theory naming for
more intuitive names for all of the algebraic structures. I also dropped almost
all of the intermediate structures such as Chain, Apply, Extend, etc. Structures
that I never abstracted over have been removed. This includes Semigroupoid,
Category, Extend, Comonad. Some of these may return in future versions if there
is demand or I find a solid use case.

In short, this commit is paradoxical. On the one hand all of the naming and
simplification changes were made to make fun more approachable and intuitive. On
the other hand I find that I have less time for programming that ever before so
interacting with these supposed new users is unlikely.

That said, these changes have been in the backlog for awhile and are finally up.
With the exception of some deno misfires on coverage reporting I've pushed test
coverage to 100%.

The next bit of work is to go through all of the algebraic data structures and
fill any features holes where an algebraic structure can be implemented or a
getter for an algebraic structure can be implemented.

FossilOrigin-Name: f46fc828a564c727414299c786ee8192e6734b59eba36226e091cc830edef291