feat: reorganize contrib, implement stream, simplify optic
This is a fairly large commit. The primary goals were to recreate the dux state
management library on top of fun. This led down a rabbit whole that included:

1. Implementing and writing tests for DatumEither
2. Exploring various stream implementations, including mostjs, rxjs, most2, and
   callbags.
3. Implementing a Stream type in fun.
4. Implementing Dux on top of Stream and DatumEither

In addition to adding Stream, Dux, and DatumEither I settled on a simpler model
for "contributions". Previous external contributions lived in the contrib
directory. These were anything from experimental ideas to algebraic structure
wrappers around other libraries. Ultimately, the purpose of fun is to be
dependency free, so external libraries were removed and "ideas" were either
promoted to full fledged modules or moved into the "ideas" directory. This
simplifies fun by making "ideas" non-production.

There is still some overlap between ideas and modules marked as experimental.
The basic idea is that modules in the ideas directoey  may or may not be published,
but modules in the root directory are definitely published. Experimental tagged
modules indicate that the api for those modules is considered unstable.

In contrib, the mostjs and fast-check modules were removed. Free was promoted to
a root module, and dux was moved to ideas. There were a handful of small api
additions to various modules such as promise, array, and async_iterable.

Lastly, the Optic type in optic.ts was simplified. Previously, Optic was broken
into an intersection of Viewer and  Modifier, with Reviewer being a separate
type. I found a way to properly do an optional include of reviewer in the root
Optic type so the subtypes of Viewer, Modifier, and Reviewer were removed. The
benefit of this is that the standard compose function in optic.ts will properly
compose a review function if it exists on the optics being composed.

may or may not be puli