feat: initial stab at parser idea
Added a very basic recursive descent parser in ideas. It's pretty
straightforward and reuses the DecodeError from Decoder for tracking
errors. In the long run, when I learn a bit more about parsers, lexers,
and grammars, I'll likely completely change the implementation to
something like an LR(1) parser generator. For now, this works ok for the
minimal places I need some basic string parsing and don't want to deal
with regex.

Removed that silly bimap function from bimappable.ts. I think the idea
was that once could create a Bimappable instance from map and
mapSecond.. But that is easier with plain types. Anyway, that's gone
now, just write your own bimap function.

Also, cleaned up a bunch of the example code that was erroring during
deno test --doc. It's not 100% yet and it seems the newest deno versions
broke my coverage script.. but that's work for another day.