feat(v2.3.4): elevated parser.ts to main library
This commit moves parser.ts to a fully supported fun module. Tests and
documentation, written by AI and reviewed by human, are at 100% and look
good. There will be a change to the error type of parser at some point,
but for now the errors are "good enough".

RELEASE: v2.3.4