feat: add Effect type to fun
Using Effect in pick lately and it's working nicely there. It's a sort
of collapsed version of AsyncStateEither. This version has all the
basics but it could benefit from some additional typeclass instance
definitions.

I've also updated the scripts/exports.sh script to output the files in
ideas as well, so I don't have to manually remember those.