feat: make Effect variadic
I've found a need in the pick library for the Effect type to be
variadic. This will resolve a lot of uncomfortable apis that the unary
type of Effect used to have. It just so happens that the output of a
pair (without using Pair) was fortuitous, since TypeScript allows for a
pretty comprehensive type when using spread operations.

Tests were also updated (with surprisingly few breaking changes). I'm
going to take this opportunity to start in on a release candidate for
version 3 of fun, since there are breaking changes in effect. Over the
next six months to a year my plan is to lazily look through the other
modules of fun for places where I want to introduce other breaking
changes. Prime candidates are:

* Schemable needs some tooling around min/max and other useful
  refinement types.
* Stream likely has issues that I've overlooked.