feat(3.0.0-rc.3): document and test effect.ts#mapEither
This change adds documentation and testint to mapEither in the effect.ts
module. Additionally, a change was mad to effect.ts#swap to use
either.ts#swap for the inner changes.

Lastly, the ./scripts/coverage.sh script was moved to flake.nix,
dependencies on darkhttpd and jq were removed, the coverage script now
outputs only a summary to the cli (this may change), and the script can
be run by calling `coverage`. I think in the future I'll move the script
back to the ./scripts directory but source it in flake.nix. I'll
seperate scripts for testing, doc testing, formatting, and such, and
compose them properly, allowing the coverage directory to stay between
runs but be ignored by git.