feat(3.0.0-rc.5): implement tagged error type in err.ts
This commit primarily implements the Err type in err.ts. It is intended
to standardize error types such that errors from different computations
can be type unioned instead of deconstructed into application errors.
The secondary tag allows for some intelligent mapping at the end of the
pipeline, ideally useful for perhaps turning different tagged errors
into http status codes, etc. Originally I had implemented Combinable for
Err as a Free combinable, but I found that this would complicate the
structure of errors. Instead I am opting to allow the user to either
wrap the Err type in their own structure or to embed errors in the typed
context field. If I find that this was the wrong direction I hope to fix
it before the 3.0.0 release.